dangdhph20582_gd

+++ src/main/java/com/example/demo/repository/TheLoaiRepository.java
@@ -4,5 +4,5 @@ import com.example.demo.model.TheLoai;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 @Repository
-public interface TheLoaiRepository extends JpaRepository<TheLoai,Integer> {
+public interface TheLoaiRepository extends JpaRepository<TheLoai,Long> {
 }
